# swagger.yaml
swagger: '2.0'
info:
  title: Furniture Rental API
  version: 1.0.0
  description: API для аренды мебели
host: localhost:3000
basePath: /
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT токен, например: `Bearer <token>`"

paths:
  /login:
    post:
      summary: Авторизация администратора
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
      responses:
        '200':
          description: Успешный вход
        '401':
          description: Неверные данные
  /api/products:
    get:
      summary: Получить список товаров
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
  /api/orders:
    get:
      summary: Получить список заказов
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
    post:
      summary: Создать новый заказ
      security:
        - Bearer: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/NewOrder'
      responses:
        '201':
          description: Заказ создан
        '400':
          description: Неверные данные
        '401':
          description: Неавторизован

definitions:
  Product:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      price:
        type: number
      image_url:
        type: string
      category:
        type: string
    required:
      - id
      - name
      - description
      - price
      - image_url

  NewOrder:
    type: object
    properties:
      items:
        type: array
        items:
          type: object
          properties:
            product_id:
              type: integer
            quantity:
              type: integer
    required:
      - items